pip_requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
dataclasses==0.8
decorator==5.1.0
docker-pycreds==0.4.0
entrypoints==0.3
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.2
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kafka-python==2.0.2
Markdown==3.3.5
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
packaging==21.2
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-mnist==0.7
pytorch-lightning==1.5.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.5.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typed-ast==1.4.3
typing_extensions==4.0.0
urllib3==1.26.7
wandb==0.12.6
wcwidth==0.2.5
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
